compile ncplot as c++ generic #446
